Open Discourse Data Object Class

Data object class. Provides functionality like retrieving data or summary functions.

Public fields

table_name

(character) character string of the table name

data

(dataframe) data frame of the retrieved table

Methods

Public methods

OData$new()

OData$get_table_data()

OData$get_con()

OData$clone()

Method new()

constructor

Usage

OData$new(table_name = "factions")

Arguments

table_name

(character) character string of the table name Get Data Table

Method get_table_data()

Retrieve Data from OD DB and keep it as data

Usage

OData$get_table_data(table_name = self$table_name)

Arguments

table_name

(character) character string of the table name Create db connection from OD docker defaults

Method get_con()

Usage

OData$get_con(
  db = "next",
  host_db = "localhost",
  db_port = "5432",
  db_user = "postgres",
  db_password = "postgres"
)

Arguments

db

(character character string of the connection arg)

host_db

(character character string of the connection arg)

db_port

(character character string of the connection arg)

db_user

(character character string of the connection arg)

db_password

(character character string of the connection arg)

Returns

RPostgres conn object

Method clone()

The objects of this class are cloneable with this method.

Usage

OData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.